Airflow/requirements.txt
apache-airflow==2.10.2
beautifulsoup4==4.12.3
pandas==2.2.3
selenium==4.25.0